+++ src/Hacks/FreezeAttempts.cpp
@@ -1,0 +1,5 @@
+#include <Geode/Geode.hpp>
+#include <Geode/modify/GameObject.hpp>
+#include "../Client/Client.h"
+
+using namespace geode::prelude;