+++ src/Labels/LabelCommon.hpp
@@ -30,7 +30,7 @@ struct LabelEvent
 {
     enum LabelEventType type;
     bool enabled = false;
-    cocos2d::ccColor3B colour = cocos2d::ccc3(0, 255, 0);
+    cocos2d::ccColor4B colour = cocos2d::ccc4(0, 255, 0, 255);
     float fadeIn = 0;
     float hold = 0;
     float fadeOut = 0.5f;