+++ src/Labels/RunInfo.cpp
@@ -1,0 +1,8 @@
+#include "RunInfo.hpp"
+
+void RunPlayLayer::resetLevel()
+{
+    PlayLayer::resetLevel();
+
+    m_fields->fromPercent = getCurrentPercent();
+}